requirements.txt
Django==4.2.17
djangorestframework==3.14.0
django-cors-headers==4.3.1
stripe==7.11.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0